fix: separate debugger
