+++ perception/multi_object_tracker/include/multi_object_tracker/multi_object_tracker_core.hpp
@@ -20,12 +20,10 @@
 #define MULTI_OBJECT_TRACKER__MULTI_OBJECT_TRACKER_CORE_HPP_
 
 #include "multi_object_tracker/data_association/data_association.hpp"
+#include "multi_object_tracker/debugger.hpp"
 #include "multi_object_tracker/tracker/model/tracker_base.hpp"
 
 #include <rclcpp/rclcpp.hpp>
-#include <tier4_autoware_utils/ros/debug_publisher.hpp>
-#include <tier4_autoware_utils/ros/published_time_publisher.hpp>
-#include <tier4_autoware_utils/system/stop_watch.hpp>
 
 #include <autoware_auto_perception_msgs/msg/detected_objects.hpp>
 #include <autoware_auto_perception_msgs/msg/tracked_objects.hpp>
@@ -41,9 +39,6 @@
 #include <tf2_geometry_msgs/tf2_geometry_msgs.hpp>
 #endif
 
-#include <diagnostic_updater/diagnostic_updater.hpp>
-#include <diagnostic_updater/publisher.hpp>
-
 #include <tf2_ros/buffer.h>
 #include <tf2_ros/transform_listener.h>
 
@@ -52,43 +47,6 @@
 #include <memory>
 #include <string>
 #include <vector>
-
-/**
- * @brief Debugger class for multi object tracker
- * @details This class is used to publish debug information of multi object tracker
- */
-class TrackerDebugger
-{
-public:
-  explicit TrackerDebugger(rclcpp::Node & node);
-  void publishTentativeObjects(
-    const autoware_auto_perception_msgs::msg::TrackedObjects & tentative_objects) const;
-  void startMeasurementTime(
-    const rclcpp::Time & now, const rclcpp::Time & measurement_header_stamp);
-  void endPublishTime(const rclcpp::Time & now, const rclcpp::Time & object_time);
-  void setupDiagnostics();
-  void checkDelay(diagnostic_updater::DiagnosticStatusWrapper & stat);
-  struct DEBUG_SETTINGS
-  {
-    bool publish_processing_time;
-    bool publish_tentative_objects;
-    double diagnostics_warn_delay;
-    double diagnostics_error_delay;
-  } debug_settings_;
-  double elapsed_time_from_sensor_input_ = 0.0;
-  diagnostic_updater::Updater diagnostic_updater_;
-
-private:
-  void loadParameters();
-  rclcpp::Node & node_;
-  rclcpp::Publisher<autoware_auto_perception_msgs::msg::TrackedObjects>::SharedPtr
-    debug_tentative_objects_pub_;
-  std::unique_ptr<tier4_autoware_utils::StopWatch<std::chrono::milliseconds>> stop_watch_ptr_;
-  std::unique_ptr<tier4_autoware_utils::DebugPublisher> processing_time_publisher_;
-  rclcpp::Time last_input_stamp_;
-  rclcpp::Time stamp_process_start_;
-  rclcpp::Time stamp_publish_output_;
-};
 
 class MultiObjectTracker : public rclcpp::Node
 {
